requirements.txt
analytics-mcp==0.1.1
google-analytics-data==0.18.19
google-analytics-admin==0.25.0
google-auth~=2.40
mcp[cli]>=1.2.0
httpx>=0.28.1
fastapi>=0.104.0
uvicorn>=0.23.0
